requirements.txt
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
albumentations>=1.3.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-image>=0.19.0
Pillow>=9.0.0
pycocotools>=2.0.6
tqdm>=4.64.0
